update parquet.net package to 4.5.4
